Modified login method from API
Modified the login method for the api such that it now returns a LoggedInUser model object with the user's id and token.

Also modified token such that it includes now the user's ID and added a SP for getting the user's ID via a given email to achieve all of this.