Completely changed the Pipelines logic
Completely changed the Pipelines logic such that it uses now a TabControl to switch between the Stages.

Created Stages View and ViewModel and wired everything up such that when a tab is selected the appropiate event is fired, the only thing left is to detect which tab has been selected.